local协议，host支持root与profile字段访问对应路径；优化代码结构目录路径
