+++ src/preload.js
@@ -1,22 +1,4 @@
-const { ipcRenderer, contextBridge } = require("electron");
-
-
-// LiteLoader
-Object.defineProperty(globalThis, "LiteLoader", {
-    value: {
-        ...ipcRenderer.sendSync("LiteLoader.LiteLoader.LiteLoader"),
-        api: {
-            config: {
-                get: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "config", "get", ...args),
-                set: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "config", "set", ...args)
-            },
-            openExternal: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "openExternal", "openExternal", ...args),
-            openPath: (...args) => ipcRenderer.invoke("LiteLoader.LiteLoader.api", "openPath", "openPath", ...args)
-        }
-    }
-});
-
-contextBridge.exposeInMainWorld("LiteLoader", LiteLoader);
+const { ipcRenderer } = require("electron");
 
 
 // 加载渲染进程
@@ -24,50 +6,38 @@ window.addEventListener("DOMContentLoaded", () => {
     const script = document.createElement("script");
     script.type = "module";
     script.defer = true;
-    script.src = `local:///${LiteLoader.path.root}/src/renderer.js`;
+    script.src = `local://root/src/renderer.js`;
     document.head.append(script);
 });
 
 
-const runPreloadScript = code => binding.createPreloadScript(`(
-    async function(
-        require,
-        process,
-        Buffer,
-        global,
-        setImmediate,
-        clearImmediate,
-        exports,
-        module
-    ) {
-        ${code}
-    }
-)`)(...arguments);
-
-
-fetch(`local:///${LiteLoader.path.root}/src/setting/preload.js`).then(async res => {
-    runPreloadScript(await res.text());
+const runPreloadScript = code => binding.createPreloadScript(`
+(async function(require, process, Buffer, global, setImmediate, clearImmediate, exports, module) {
+    ${code}
 });
+`)(...arguments);
 
 
-// 加载插件 Preload
-for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
-    if (plugin.disabled || plugin.incompatible) {
-        continue;
+(async () => {
+    runPreloadScript(await (await fetch(`local://root/src/liteloader_api/preload.js`)).text());
+    runPreloadScript(await (await fetch(`local://root/src/setting/preload.js`)).text());
+    // 加载插件 Preload
+    for (const [slug, plugin] of Object.entries(LiteLoader.plugins)) {
+        if (plugin.disabled || plugin.incompatible) {
+            continue;
+        }
+        if (plugin.path.injects.preload) {
+            runPreloadScript(await (await fetch(`local:///${plugin.path.injects.preload}`)).text());
+        }
     }
-    if (plugin.path.injects.preload) {
-        fetch(`local:///${plugin.path.injects.preload}`).then(async res => {
-            runPreloadScript(await res.text());
-        });
-    }
-}
+})();
 
 
 // 加载 QQNT Preload
-let flag = false;
+let isLoaded = false;
 ipcRenderer.invoke("LiteLoader.LiteLoader.preload").then(preload => {
-    if (!flag) {
+    if (!isLoaded) {
         runPreloadScript(preload);
-        flag = true;
+        isLoaded = true;
     }
 });
